Debug but count dosen't work

+++ ft_putstr.c
@@ -6,7 +6,7 @@
 /*   By: bfaure <bfaure@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/08 14:36:17 by bfaure            #+#    #+#             */
-/*   Updated: 2022/12/08 14:58:36 by bfaure           ###   ########lyon.fr   */
+/*   Updated: 2022/12/12 16:11:54 by bfaure           ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ static size_t	ft_strlen(char *str)
 	return (i);
 }
 
-size_t	ft_putstr(char *str, size_t size)
+int	ft_putstr(char *str, int size)
 {
 	write(1, str, ft_strlen(str));
 	return (size + ft_strlen(str));
